requirements.txt
aiohappyeyeballs==2.3.5
aiohttp==3.10.2
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.4.0
appnope==0.1.4
asttokens==2.4.1
attrs==24.2.0
bitarray==2.9.2
certifi==2024.7.4
charset-normalizer==3.3.2
ckzg==1.0.2
comm==0.2.2
cytoolz==0.12.3
debugpy==1.8.5
decorator==5.1.1
dnspython==2.6.1
eth-account==0.13.1
eth-hash==0.7.0
eth-keyfile==0.8.1
eth-keys==0.5.1
eth-rlp==2.1.0
eth-typing==5.0.0b3
eth-utils==5.0.0b1
eth_abi==5.1.0
executing==2.0.1
frozenlist==1.4.1
h11==0.14.0
hexbytes==1.2.1
httpcore==1.0.5
httpx==0.27.0
idna==3.7
jedi==0.19.1
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
jupyter_client==8.6.2
jupyter_core==5.7.2
lru-dict==1.2.0
matplotlib-inline==0.1.7
mongopersistence==0.3.1
motor==3.5.1
multidict==6.0.5
nest-asyncio==1.6.0
packaging==24.1
parsimonious==0.10.0
parso==0.8.4
pexpect==4.9.0
platformdirs==4.2.2
prompt_toolkit==3.0.47
protobuf==5.27.3
psutil==6.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
pycryptodome==3.20.0
pydantic==2.8.2
pydantic_core==2.20.1
Pygments==2.18.0
pymongo==4.8.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-telegram-bot==21.4
pyunormalize==15.1.0
pyzmq==26.1.0
referencing==0.35.1
regex==2024.7.24
requests==2.32.3
rlp==4.0.1
rpds-py==0.20.0
six==1.16.0
sniffio==1.3.1
stack-data==0.6.3
tk==0.1.0
toolz==0.12.1
tornado==6.4.1
traitlets==5.14.3
types-requests==2.32.0.20240712
typing_extensions==4.12.2
urllib3==2.2.2
wcwidth==0.2.13
web3==7.0.0b9
websockets==12.0
yarl==1.9.4
